distribution: move bootc functions to relevant files
The "distroregistry.go" file contains everything operates on the
collection of distributions. The "distribution.go" contains operations
on a single entry.
